refactor LabeledRadioButtonGroup component
