+++ packages/react-components/src/molecules/LabeledRadioButtonGroup.tsx
@@ -1,12 +1,28 @@
+import React, { useRef } from 'react';
+import { v4 as uuidV4 } from 'uuid';
 import { css } from '@emotion/react';
 
-import { RadioButtonGroup, RadioButtonGroupProps } from '../organisms';
-import { perRem } from '../pixels';
+import { Option } from '../select';
+import { noop } from '../utils';
+import { LabeledRadioButton } from '.';
+import { mobileScreen, perRem } from '../pixels';
 
-type LabeledRadioButtonGroupProps<V extends string> = {
-  readonly title?: React.ReactNode;
-  readonly subtitle?: React.ReactNode;
-} & RadioButtonGroupProps<V>;
+export interface LabeledRadioButtonGroupProps<V extends string> {
+  readonly title?: string;
+  readonly subtitle?: string;
+  readonly options: ReadonlyArray<Option<V>>;
+
+  readonly value: V;
+  readonly onChange?: (newValue: V) => void;
+}
+
+const optionListStyles = css({
+  display: 'grid',
+  gridTemplateColumns: '1fr 1fr 1fr',
+  [`@media (max-width: ${mobileScreen.min}px)`]: {
+    gridTemplateColumns: '1fr',
+  },
+});
 
 const containerStyles = css({
   border: 'none',
@@ -21,15 +37,31 @@ const subtitleStyles = css({
 export default function LabeledRadioButtonGroup<V extends string>({
   title,
   subtitle,
-  ...radioButtonGroupProps
+  options,
+
+  value,
+  onChange = noop,
 }: LabeledRadioButtonGroupProps<V>): ReturnType<React.FC> {
+  const groupName = useRef(uuidV4());
   return (
     <fieldset css={containerStyles}>
-      <legend>
-        <strong>{title}</strong>
-        <span css={subtitleStyles}>{subtitle}</span>
-      </legend>
-      <RadioButtonGroup {...radioButtonGroupProps} />
+      {title || subtitle ? (
+        <legend>
+          <strong>{title}</strong>
+          <span css={subtitleStyles}>{subtitle}</span>
+        </legend>
+      ) : null}
+      <div css={optionListStyles}>
+        {options.map((option) => (
+          <LabeledRadioButton
+            key={option.value}
+            groupName={groupName.current}
+            title={option.label}
+            checked={option.value === value}
+            onSelect={() => onChange(option.value)}
+          />
+        ))}
+      </div>
     </fieldset>
   );
 }
