+++ packages/react-components/src/organisms/index.ts
@@ -30,7 +30,6 @@ export { default as PeopleCard } from './PeopleCard';
 export { default as ProfileCardList } from './ProfileCardList';
 export { default as ProfileExpertiseAndResources } from './ProfileExpertiseAndResources';
 export { default as QuestionsSection } from './QuestionsSection';
-export { default as RadioButtonGroup } from './RadioButtonGroup';
 export { default as ResultList } from './ResultList';
 export { default as RichText } from './RichText';
 export { default as RichTextCard } from './RichTextCard';
@@ -58,5 +57,3 @@ export { default as UserProfilePlaceholderCard } from './UserProfilePlaceholderC
 export { default as UserProfileRecentWorks } from './UserProfileRecentWorks';
 export { default as UserProfileRole } from './UserProfileRole';
 export { default as WelcomeCard } from './WelcomeCard';
-
-export type { RadioButtonGroupProps } from './RadioButtonGroup';
